+++ core/lib/compass/core/sass_extensions/functions/inline_image.rb
@@ -2,7 +2,7 @@ module Compass::Core::SassExtensions::Functions::InlineImage
 
   def inline_image(path, mime_type = nil)
     path = path.value
-    real_path = File.join(Compass.configuration.images_path, path)
+    real_path = Compass.configuration.url_resolver.find_asset(:image, path)
     inline_image_string(data(real_path), compute_mime_type(path, mime_type))
   end
 
@@ -10,7 +10,7 @@ module Compass::Core::SassExtensions::Functions::InlineImage
     files = []
     with_each_font_file(*args) do |path, type|
       path = path.value
-      real_path = File.join(Compass.configuration.fonts_path, path)
+      real_path = Compass.configuration.url_resolver.find_asset(:font, path)
       data = inline_image_string(data(real_path), compute_mime_type(path))
       files << list(data, unquoted_string("format('#{type}')"), :space)
     end